refactor: reorganize dir and renamed components
