+++ seankwarren/react-flow-calc/src/components/reusable/DraggableButton.jsx
@@ -1,9 +1,9 @@
 import React from 'react';
 
-const DraggableButton = ({ label, nodeType, onDragStart }) => {
+const DraggableButton = ({ label, nodeType, onDragStart, className }) => {
     return (
         <button
-            className={`node button ${nodeType}`}
+            className={`node button ${nodeType} ${className}`}
             draggable='true'
             onDragStart={(e) => onDragStart(e, nodeType)}
         >
